Projects, Auth, Container Registry, Stop Invocation (#7)
- Add `Project` resource
- Add authentication and authorization system
- Update `Function` schema
- Update `Invocation` schema
- Add "stop Invocation" route
- Use Views instead of Mango queries
- Expose Container Registry routes

+++ src/api/functions/v1/deleteFunction.ts
@@ -1,10 +1,10 @@
 import type { RouteOptions } from '@brer/fastify'
 import S from 'fluent-json-schema-es'
 
-import { getFunctionId } from '../../../lib/function.js'
+import { getFunctionByName } from '../../../lib/function.js'
 import { getLabelSelector } from '../../../lib/kubernetes.js'
 
-interface RouteGeneric {
+export interface RouteGeneric {
   Params: {
     functionName: string
   }
@@ -29,15 +29,17 @@ export default (): RouteOptions<RouteGeneric> => ({
     },
   },
   async handler(request, reply) {
-    const { database, kubernetes, tasks } = this
-    const { params } = request
+    const { auth, kubernetes, store, tasks } = this
+    const { params, session } = request
 
-    const fn = await database.functions
-      .find(getFunctionId(params.functionName))
-      .unwrap()
-
+    const fn = await getFunctionByName(store, params.functionName)
     if (!fn) {
       return reply.code(404).error({ message: 'Function not found.' })
+    }
+
+    const result = await auth.authorize(session, 'viewer', fn.project)
+    if (result.isErr) {
+      return reply.error(result.unwrapErr())
     }
 
     await kubernetes.api.CoreV1Api.deleteCollectionNamespacedPod(
@@ -50,13 +52,21 @@ export default (): RouteOptions<RouteGeneric> => ({
       getLabelSelector({ functionName: fn.name }),
     )
 
-    await database.functions.from(fn).delete().consume()
+    await store.functions.from(fn).delete().consume({ purge: true })
 
     tasks.push(async log => {
-      const count = await database.invocations
-        .filter({ functionName: fn.name })
+      const count = await store.invocations
+        .filter({
+          _design: 'default',
+          _view: 'by_project',
+          startkey: [fn.project, fn.name, null],
+          endkey: [fn.project, fn.name, {}],
+        })
         .delete()
-        .consume()
+        .consume({
+          purge: true,
+          sorted: false,
+        })
 
       log.debug(`deleted ${count} ${fn.name} invocation(s)`)
     })
